Correct __import__ level argument for python 3.x
Python 3 doesn't allow level to be -1 anymore. One has to specify whether to use absolute search path or how much relative parent directory must be searched.
Using -1 results in a ValueError.
Value 0 was tested on both python 3.3 and 2.7 when importing the guessit module from the site-package install or from a relative folder.

Another solution would be not to specify the level argument, as it is -1 by default on python 2 and 0 on python 3, which should work correctly (untested).