Using badges from shields.io
As badges from `pypip.in` are down for some time now, updated `README` to use badges from [shields.io](http://shields.io/) instead. They have the added benefit of being in `SVG` which make them look good on higher resolutions.

One drawback is that they don't have a badge for licensing, so I removed that badge from `README` as that information appears at the bottom of the document as well.