feat: add the input filename to the output as 'raw_input'
Passing the input filename through to the output allows consumers of the
`guessit` command output to grab the filename directly in the same way
as other properties. This is especially helpful when processing multiple
filenames and iterating over the results.

The `raw_input` key was only added when running the CLI tool since code
that uses `guessit` as a library won't need this functionality.